energy: [WIP] implement libsensors-based monitoring
This commit starts using libsensors to consume HWMON information.
I opted for the C library because it handles scaling some units and
other hardware quirks automatically, though the API is pretty confusing
and has no examples to speak of.

This commit hijacks the main function from its original purpose of
generating CSV data. I intend to return to that soon.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>