watt-wiser: split into several files and delay heading load
This commit moves chart logic into chart.go, line reader logic into linereader.go,
and restructures the CSV read process to send headers to the UI over the output
channel instead of needing them available in advance.

This refactor is in service of allowing the UI to actually start the sensors itself,
rather than requiring them to already be running.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>