rapl: add primitive windows support
This commit contains a port of the userspace code for accessing scaphandre's RAPL MSR
windows driver. You must already have the windows driver installed for this to work.
Additionally, this fetches only the total energy counter for the first package, as
the upstream code I used as a reference only supported that. We should try to fetch
the more granular energy counters soon.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>