backend,main: rework session management
Data is now managed in sessions, which can either be live measurements or
replays of recorded data. Benchmarks can be created against live measurements,
and the benchmark file name identifies the corresponding session. In future
commits, this will be used to allow us to load multiple benchmarks from different
sessions in parallel and compare them.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>