main(chart): fix infinite memory leak from NaN
This commit fixes an incredibly subtle memory leak. If the data for the chart initially has a
maximum range rate of zero, it can introduce a NaN into the computation to determine the Y axis
scale of the graph. This NaN then makes the loop that draws the Y axis grid go into an infinite
loop of drawing rectangles, eventually exhausting system memory.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>