backend: make csv and benchmark file creation more reliable
Apparently opening a file in the file explorer changes the working directory of our process.
This can trivially make us try to create our output files in random locations throughout disk,
and we get permission errors for many places. Instead, we should just place files relative to
the executable and move on with our lives.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>